% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apache2.R
\name{age_score}
\alias{age_score}
\title{Calculate APACHE II Age Score}
\usage{
age_score(x)
}
\arguments{
\item{x}{A numeric vector of ages}
}
\description{
\code{age_score} calculates the age score for the APACHE II
}
\details{
This function calculates the Age Score based on the APACHE II scoring system.
}
\examples{

}
\keyword{internal}

