% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apache3.R
\name{apache3_neuro}
\alias{apache3_neuro}
\title{Calculate neurologic abnormality score for APACHE III}
\usage{
apache3_neuro(eye, motor, verbal)
}
\arguments{
\item{eye}{A numeric vector}

\item{motor}{A numeric vector}

\item{verbal}{A numeric vector}
}
\value{
An integer vector
}
\description{
Calculate neurologic abnormality score for APACHE III
}

