% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apache3.R
\name{apache3_score}
\alias{apache3_score}
\alias{apache3_score.aa_grad}
\alias{apache3_score.age}
\alias{apache3_score.albumin}
\alias{apache3_score.bili}
\alias{apache3_score.bun}
\alias{apache3_score.default}
\alias{apache3_score.glucose}
\alias{apache3_score.hct}
\alias{apache3_score.hr}
\alias{apache3_score.map}
\alias{apache3_score.pao2}
\alias{apache3_score.ph}
\alias{apache3_score.rr}
\alias{apache3_score.scr}
\alias{apache3_score.sodium}
\alias{apache3_score.temp}
\alias{apache3_score.uop}
\alias{apache3_score.wbc}
\title{Calculate APACHE III Acute Physiologic Score}
\usage{
apache3_score(x, ...)

\method{apache3_score}{default}(x, ...)

\method{apache3_score}{temp}(x, ...)

\method{apache3_score}{map}(x, ...)

\method{apache3_score}{hr}(x, ...)

\method{apache3_score}{rr}(x, ...)

\method{apache3_score}{ph}(x, ..., pco2)

\method{apache3_score}{sodium}(x, ...)

\method{apache3_score}{scr}(x, ..., arf)

\method{apache3_score}{uop}(x, ...)

\method{apache3_score}{bun}(x, ...)

\method{apache3_score}{hct}(x, ...)

\method{apache3_score}{wbc}(x, ...)

\method{apache3_score}{bili}(x, ...)

\method{apache3_score}{albumin}(x, ...)

\method{apache3_score}{glucose}(x, ...)

\method{apache3_score}{pao2}(x, ...)

\method{apache3_score}{aa_grad}(x, ...)

\method{apache3_score}{age}(x, ...)
}
\arguments{
\item{x}{A numeric vector with an icuriskr class type}

\item{...}{additional arguments passed on to individual methods}
}
\description{
\code{apache3_score} calculates the APS score for an APACHE III variable
}
\details{
This is an S3 generic function for calculating the Acute Physicologic Score
(APS) for a variable based on the APACHE III scoring system.The function
invokes the appropriate method based on the type of data (i.e., temperature,
mean arterial pressure, etc.).
}
\examples{

}

