% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{aps2}
\alias{aps2}
\alias{as.aps2}
\title{Construct Acute Physiology Score data types for APACHE II}
\usage{
aps2(x)

as.aps2(x)
}
\arguments{
\item{x}{object to set class}
}
\description{
Takes an R object and sets class to an aps type.
}
\keyword{internal}

