% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{aps3}
\alias{aps3}
\alias{as.aps3}
\title{Construct Acute Physiology Score data types for APACHE III}
\usage{
aps3(x)

as.aps3(x)
}
\arguments{
\item{x}{object to set class}
}
\description{
Takes an R object and sets class to an aps type.
}
\keyword{internal}

