% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apache2.R
\name{aps_score}
\alias{aps_score}
\alias{aps_score.default}
\alias{aps_score.gcs}
\alias{aps_score.hco3}
\alias{aps_score.hct}
\alias{aps_score.hr}
\alias{aps_score.map}
\alias{aps_score.pao2}
\alias{aps_score.ph}
\alias{aps_score.potassium}
\alias{aps_score.rr}
\alias{aps_score.scr}
\alias{aps_score.sodium}
\alias{aps_score.temp}
\alias{aps_score.wbc}
\title{Calculate APACHE II Acute Physiologic Score}
\usage{
aps_score(x, ...)

\method{aps_score}{default}(x, ...)

\method{aps_score}{temp}(x, ...)

\method{aps_score}{map}(x, ...)

\method{aps_score}{hr}(x, ...)

\method{aps_score}{rr}(x, ...)

\method{aps_score}{ph}(x, ...)

\method{aps_score}{sodium}(x, ...)

\method{aps_score}{potassium}(x, ...)

\method{aps_score}{scr}(x, ...)

\method{aps_score}{hct}(x, ...)

\method{aps_score}{wbc}(x, ...)

\method{aps_score}{gcs}(x, ...)

\method{aps_score}{hco3}(x, ...)

\method{aps_score}{pao2}(x, ...)
}
\arguments{
\item{x}{A numeric vector with an icuriskr class type}

\item{...}{additional arguments passed on to individual methods}
}
\description{
\code{aps_score} calculates the APS score for an APACHE II variable
}
\details{
This is an S3 generic function for calculating the Acute Physicologic Score
(APS) for a variable based on the APACHE II scoring system.The function
invokes the appropriate method based on the type of data (i.e., temperature,
mean arterial pressure, etc.).
}
\examples{

}
\keyword{internal}

