% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidity_drg.R
\docType{data}
\name{comorbidity_drg}
\alias{comorbidity_drg}
\title{Comorbidity mapping using ICD-9-CM}
\format{A data frame with comorbidity and DRG MDC codes.
\describe{
  \item{comorbidity}{name of the comorbidity}
  \item{drg}{the DRG Major Disease Classification (MDC) code}
}}
\usage{
comorbidity_drg
}
\description{
A dataset containing DRG MDC codes which can be used to exclude an ICD code
as a comorbidity.
}
\keyword{datasets}

