% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidity_map_icd10.R
\docType{data}
\name{comorbidity_map_icd10_elix}
\alias{comorbidity_map_icd10_elix}
\title{Comorbidity mapping using ICD-10-CM}
\format{A named list with a character vector of ICD-10-CM codes which
  correspond to the comorbidities.}
\usage{
comorbidity_map_icd10_elix
}
\description{
A list containing the ICD-10-CM codes for comorbidities included in APAPCHE
II, APAPCHE III, and SAPS II. Based on original Elixhauser list.
}
\keyword{datasets}

