% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidity_map_icd9.R
\docType{data}
\name{comorbidity_map_icd9}
\alias{comorbidity_map_icd9}
\title{Comorbidity mapping using ICD-9-CM}
\format{A named list with a character vector of ICD-9-CM codes which
  correspond to the comorbidities.}
\usage{
comorbidity_map_icd9
}
\description{
A list containing the ICD-9-CM codes for comorbidities included in APAPCHE
II, APAPCHE III, and SAPS II.
}
\keyword{datasets}

