% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{F_to_C}
\alias{F_to_C}
\title{Convert from Farenheit to Celsius}
\usage{
F_to_C(degree)
}
\arguments{
\item{degree}{A numeric, the temperature in degrees Farenheit}
}
\value{
A numeric, the temperature in degrees Celsius
}
\description{
This function converts temperatures from degrees Farenheit to degrees
Celsius.
}

