% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icuriskr.R
\docType{package}
\name{icuriskr}
\alias{icuriskr}
\alias{icuriskr-package}
\title{Calculate ICU Morbidity and Mortality Risk Scores}
\description{
\pkg{icuriskr} provides methods to calculate ICU morbidity and mortality risk
scores
}
\details{
This package calculates different ICU morbidity and mortality risk scores,
including APACHE II, APAPCHE III, and SAPS II. The presence of comorbidities
is determined by using ICD-9-CM and ICD-10-CM codes.
}

