% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is.aps2}
\alias{is.aps2}
\title{Test icuriskr-related classes}
\usage{
is.aps2(x)
}
\arguments{
\item{x}{object which may have a aps2 class type}
}
\description{
Takes an R object and checks for an aps2 class type (APACHE II).
}
\keyword{internal}

