% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is.aps3}
\alias{is.aps3}
\title{Test icuriskr-related classes}
\usage{
is.aps3(x)
}
\arguments{
\item{x}{object which may have a aps3 class type}
}
\description{
Takes an R object and checks for an aps3 class type (APACHE III).
}
\keyword{internal}

