% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is.physiol}
\alias{is.aa_grad}
\alias{is.admit}
\alias{is.age}
\alias{is.albumin}
\alias{is.bili}
\alias{is.bun}
\alias{is.comorbidity}
\alias{is.gcs}
\alias{is.glucose}
\alias{is.hco3}
\alias{is.hct}
\alias{is.hr}
\alias{is.map}
\alias{is.pao2}
\alias{is.ph}
\alias{is.physiol}
\alias{is.potassium}
\alias{is.rr}
\alias{is.sbp}
\alias{is.scr}
\alias{is.sodium}
\alias{is.temp}
\alias{is.uop}
\alias{is.wbc}
\title{Test icuriskr-related classes}
\usage{
is.physiol(x)

is.temp(x)

is.sbp(x)

is.map(x)

is.hr(x)

is.rr(x)

is.ph(x)

is.sodium(x)

is.potassium(x)

is.scr(x)

is.bun(x)

is.hct(x)

is.wbc(x)

is.gcs(x)

is.hco3(x)

is.pao2(x)

is.aa_grad(x)

is.bili(x)

is.albumin(x)

is.glucose(x)

is.uop(x)

is.age(x)

is.admit(x)

is.comorbidity(x)
}
\arguments{
\item{x}{object which may have a physiology class type}
}
\description{
Takes an R object and checks for a physiology class type.
}
\keyword{internal}

