% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{is.riskr}
\alias{is.gcs}
\alias{is.hco3}
\alias{is.hct}
\alias{is.hr}
\alias{is.map}
\alias{is.pao2}
\alias{is.ph}
\alias{is.potassium}
\alias{is.riskr}
\alias{is.rr}
\alias{is.scr}
\alias{is.sodium}
\alias{is.temp}
\alias{is.wbc}
\title{Test icuriskr-related classes}
\usage{
is.riskr(x)

is.temp(x)

is.map(x)

is.hr(x)

is.rr(x)

is.ph(x)

is.sodium(x)

is.potassium(x)

is.scr(x)

is.hct(x)

is.wbc(x)

is.gcs(x)

is.hco3(x)

is.pao2(x)
}
\arguments{
\item{x}{object which may have a riskr class type}
}
\description{
Takes an R object and checks for a riskr class type.
}
\keyword{internal}

