% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is.saps}
\alias{is.saps}
\title{Test icuriskr-related classes}
\usage{
is.saps(x)
}
\arguments{
\item{x}{object which may have a saps class type}
}
\description{
Takes an R object and checks for an saps class type.
}
\keyword{internal}

