% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{physiol}
\alias{as.aa_grad}
\alias{as.admit}
\alias{as.age}
\alias{as.albumin}
\alias{as.bili}
\alias{as.bun}
\alias{as.comorbidity}
\alias{as.gcs}
\alias{as.glucose}
\alias{as.hco3}
\alias{as.hct}
\alias{as.hr}
\alias{as.map}
\alias{as.pao2}
\alias{as.ph}
\alias{as.physiol}
\alias{as.potassium}
\alias{as.rr}
\alias{as.sbp}
\alias{as.scr}
\alias{as.sodium}
\alias{as.temp}
\alias{as.uop}
\alias{as.wbc}
\alias{physiol}
\title{Construct generic data types for use in risk scores}
\usage{
physiol(x)

as.physiol(x)

as.temp(x)

as.sbp(x)

as.map(x)

as.hr(x)

as.rr(x)

as.ph(x)

as.sodium(x)

as.potassium(x)

as.scr(x)

as.bun(x)

as.hct(x)

as.wbc(x)

as.gcs(x)

as.hco3(x)

as.pao2(x)

as.aa_grad(x)

as.bili(x)

as.albumin(x)

as.glucose(x)

as.uop(x)

as.age(x)

as.admit(x)

as.comorbidity(x)
}
\arguments{
\item{x}{object to set class}
}
\description{
Takes an R object and sets class to a specific physiologic type.
}
\keyword{internal}

