% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{riskr_class}
\alias{as.gcs}
\alias{as.hco3}
\alias{as.hct}
\alias{as.hr}
\alias{as.map}
\alias{as.pao2}
\alias{as.ph}
\alias{as.potassium}
\alias{as.riskr}
\alias{as.rr}
\alias{as.scr}
\alias{as.sodium}
\alias{as.temp}
\alias{as.wbc}
\alias{riskr}
\alias{riskr_class}
\title{Construct icuriskr data types}
\usage{
riskr(x)

as.riskr(x)

as.temp(x)

as.map(x)

as.hr(x)

as.rr(x)

as.ph(x)

as.sodium(x)

as.potassium(x)

as.scr(x)

as.hct(x)

as.wbc(x)

as.gcs(x)

as.hco3(x)

as.pao2(x)
}
\arguments{
\item{x}{object to set class \code{riskr}}
}
\description{
Takes an R object and sets class to an icuriskr type.
}
\keyword{internal}

