% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{saps}
\alias{as.saps}
\alias{saps}
\title{Construct Acute Physiology Score data types for SAPS II}
\usage{
saps(x)

as.saps(x)
}
\arguments{
\item{x}{object to set class}
}
\description{
Takes an R object and sets class to a saps type.
}
\keyword{internal}

