% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saps2.R
\name{saps2_score}
\alias{saps2_score}
\alias{saps2_score.admit}
\alias{saps2_score.age}
\alias{saps2_score.bili}
\alias{saps2_score.bun}
\alias{saps2_score.default}
\alias{saps2_score.gcs}
\alias{saps2_score.hco3}
\alias{saps2_score.hr}
\alias{saps2_score.pao2}
\alias{saps2_score.potassium}
\alias{saps2_score.sbp}
\alias{saps2_score.sodium}
\alias{saps2_score.temp}
\alias{saps2_score.uop}
\alias{saps2_score.wbc}
\title{Calculate SAPS II Score}
\usage{
saps2_score(x, ...)

\method{saps2_score}{default}(x, ...)

\method{saps2_score}{temp}(x, ...)

\method{saps2_score}{sbp}(x, ...)

\method{saps2_score}{hr}(x, ...)

\method{saps2_score}{sodium}(x, ...)

\method{saps2_score}{potassium}(x, ...)

\method{saps2_score}{bun}(x, ...)

\method{saps2_score}{bili}(x, ...)

\method{saps2_score}{wbc}(x, ...)

\method{saps2_score}{gcs}(x, ...)

\method{saps2_score}{hco3}(x, ...)

\method{saps2_score}{pao2}(x, ...)

\method{saps2_score}{uop}(x, ...)

\method{saps2_score}{age}(x)

\method{saps2_score}{admit}(x)
}
\arguments{
\item{x}{A numeric vector with an icuriskr class type}

\item{...}{additional arguments passed on to individual methods}
}
\description{
\code{saps2_score} calculates the saps score for an SAPS II variable
}
\details{
This is an S3 generic function for calculating the Acute Physicologic Score
for a variable based on the SAPS II scoring system.The function invokes the
appropriate method based on the type of data (i.e., temperature, systolic
arterial pressure, etc.).
}
\examples{

}

