% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_types}
\alias{set_types}
\title{Set class types}
\usage{
set_types(df, types)
}
\arguments{
\item{df}{A data frame}

\item{types}{A named list with column name and setter function (ex: list(hr =
as.hr))}
}
\value{
A data frame
}
\description{
This function takes physiologic variables in a data frame and sets them to
the appropriate class types.
}
\keyword{internal}

